Major Changes
Added the option to read data from a .csv file. Right now the code is a
little confusing to run, as you need to alternate between generating a
signal or reading from a file. Added a folder with an example of .csv
data. Fixed a mistake on freqZoom function. Now it should work correctly
for any given signal.